🚚 rename: 프로젝트명 변경
