+++ src/test/java/com/mulmeong/member/read/MemberReadApplicationTests.java
@@ -1,9 +1,9 @@
-package com.mulmeong.member_read;
+package com.mulmeong.member.read;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-@SpringBootTest
+//@SpringBootTest
 class MemberReadApplicationTests {
 
     @Test
